chore: Moved the metric submodule out of the analysis module into its own module.
