+++ hipcheck/src/analysis.rs
@@ -2,11 +2,9 @@
 
 #[allow(clippy::module_inception)]
 pub mod analysis;
-pub mod metric;
 pub mod report_builder;
 pub mod score;
 pub mod session;
 
 pub use analysis::AnalysisProvider;
 pub use analysis::AnalysisProviderStorage;
-pub use metric::MetricProvider;
