+++ hipcheck/src/metric/binary_detector/mod.rs
