+++ hipcheck/src/metric/churn.rs
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::analysis::metric::math::mean;
-use crate::analysis::metric::math::std_dev;
-use crate::analysis::MetricProvider;
 use crate::context::Context as _;
 use crate::data::git::Commit;
 use crate::error::Result;
 use crate::hc_error;
+use crate::metric::math::mean;
+use crate::metric::math::std_dev;
+use crate::metric::MetricProvider;
 use crate::TryAny;
 use crate::TryFilter;
 use crate::F64;
