+++ hipcheck/src/metric/math.rs
