+++ hipcheck/src/metric/module.rs
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::analysis::MetricProvider;
 use crate::context::Context as _;
 use crate::data::ModuleGraph;
 use crate::error::Result;
+use crate::metric::MetricProvider;
 use serde::Serialize;
 use std::rc::Rc;
