+++ hipcheck/src/metric/module_contributors.rs
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::analysis::MetricProvider;
 use crate::context::Context as _;
 use crate::data::git::Contributor;
 use crate::data::Module;
 use crate::error::Result;
+use crate::metric::MetricProvider;
 use serde::Serialize;
 use std::collections::HashMap;
 use std::rc::Rc;
